libglvnd: Update to 1.3.4
Also removed the need of wayland-egl.

Acording to
https://us.download.nvidia.com/XFree86/Linux-x86_64/495.46/README/gbm.html

egl-wayland version 1.1.8 or later must be present (if installed separately
from the the NVIDIA driver). To support applications that use the wl_drm
Wayland protocol to select a GPU to render on, such as Xwayland, version
1.1.9 or later must be present.

The egl-wayland could be installed from the NVIDIA Driver package.