nvidia: 495.44 Linux Driver Released With GBM Support
 - https://www.phoronix.com/scan.php?page=news_item&px=NVIDIA-495.44-Linux-Driver

"The NVIDIA 495.44 Linux driver is the stabilized version of the earlier
beta, that most notably introduces Generic Buffer Manager (GBM) support
for sharply improving the proprietary driver's Wayland support."

 - https://us.download.nvidia.com/XFree86/Linux-x86_64/495.46/README/gbm.html

"The following are necessary to enable use of the NVIDIA driver's GBM
backend and its EGL and related Wayland client support:

 - DRM KMS must be enabled. See Chapter 35, Direct Rendering Manager
Kernel Modesetting (DRM KMS) for details.

 - libgbm.so.1 from Mesa version 21.2 (https://gitlab.freedesktop.org/mesa/mesa).

 - egl-wayland version 1.1.8 or later must be present (if installed separately
from the the NVIDIA driver). To support applications that use the wl_drm Wayland
protocol to select a GPU to render on, such as Xwayland, version 1.1.9 or
later must be present."

Signed-off-by: Raul Muñoz <raul@foundries.io>